requirements.txt
fastapi>=0.100.0,<0.112.0
uvicorn[standard]>=0.20.0,<0.30.0
google-generativeai>=0.5.0,<0.7.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=1.10.0,<3.0.0